% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_arscape.R
\name{run_arscape}
\alias{run_arscape}
\title{Run ARscape Analysis Pipeline}
\usage{
run_arscape(
  fold_change,
  hits_fold_change = NULL,
  mock_ips = NULL,
  excluded_peptides = NULL,
  annotation_cols = c("u_pep_id", "pep_id", "pos_start", "pos_end", "UniProt_acc",
    "pep_aa", "taxon_genus", "taxon_species", "gene_symbol", "product"),
  max_iterations = 10,
  p_cutoff = 10^-4,
  score_cutoff = 2,
  min_peptides = 50,
  exclusion_method = "genus",
  progress_bar = FALSE
)
}
\arguments{
\item{fold_change}{Data frame containing peptide-level fold changes.
Must contain annotation columns (specified in \code{annotation_cols}) and sample columns.}

\item{mock_ips}{Optional vector of column names in \code{hits_fold_change} corresponding to mock IP controls. If provided along with \code{hits_fold_change},
peptides with hits in these controls are excluded.
In addition, a character vector of \code{excluded_peptides} can be provided to replace and/or supplement this peptide exclusion criterion.}

\item{excluded_peptides}{Optional character vector of UNIQUE peptide identifiers to exclude manually. These peptide identifiers MUST be of the same type as the first element in \code{annotation_cols}, typically \code{u_pep_id}.}

\item{annotation_cols}{Character vector of column names that are NOT samples.
Defaults to standard PhIP-seq annotations in the Larman Lab: \code{c("u_pep_id", "pep_id", "pos_start", "pos_end", "UniProt_acc", "pep_aa", "taxon_genus", "taxon_species", "gene_symbol", "product")}. NB! The first vector element MUST refer to the column containing UNIQUE peptide identifiers (typically \code{u_pep_id}).}

\item{max_iterations}{Integer; maximum number of iterations for the background definition algorithm.}

\item{p_cutoff}{Numeric; -log10 p-value cutoff for defining positive hits. Default is 4 (-log10(0.0001)).}

\item{score_cutoff}{Numeric; ARscore cutoff for defining positive hits. Default is 2.}

\item{min_peptides}{Integer; minimum number of peptides required per group to calculate a score. Default is 50.}

\item{exclusion_method}{Character; method for excluding reactive peptides during iteration.
Options: "genus", "species", "group".}

\item{progress_bar}{Logical; If using parallel processing, should a progress bar be displayed? No bar will be displayed if \code{future::plan(sequential)} is used, see \code{furrr::future_map()}.}
}
\value{
A data frame containing the final ARscores and p-values for all samples.
}
\description{
The main wrapper function to process PhIP-Seq fold-change data and generate
Aggregate Reactivity Scores (ARscores).
}
\section{Parallelization}{

This function leverages the \code{furrr} package for parallel processing.

To enable parallel execution:
\preformatted{
n_cores <- parallelly::availableCores()
cl <- parallel::makeForkCluster(n_cores)
future::plan(cluster, workers = cl)
}
}

